+++ app/controllers/trade_log_entries_controller.rb
@@ -1,12 +1,20 @@
-class TradeLogEntriesController < AuthenticationController
+# :nodoc:
+class TradeLogEntriesController < LogEntriesController
   skip_before_action :authenticate_user!, only: [:show]
 
   add_crumb('Home', '/')
   before_filter :load_user
   before_filter :load_character
   before_filter :load_log_entry, only: [:show, :edit, :update, :destroy]
+  before_filter :load_current_magic_items,
+                only: [:new, :create, :edit, :update]
+  before_filter :build_new_magic_item,
+                only: [:new, :create, :edit, :update]
+  before_filter :clean_up_params, only: [:create, :update]
 
-  before_filter { add_crumb @character.name, user_character_path(@character.user, @character) }
+  before_filter do
+    add_crumb @character.name, user_character_path(@character.user, @character)
+  end
 
   before_filter(only: [:new]) { add_crumb 'New Trade Log Entry' }
   before_filter(only: [:show, :edit]) { add_crumb 'Trade Log Entry' }
@@ -19,47 +27,37 @@ class TradeLogEntriesController < AuthenticationController
 
   def new
     @log_entry = @character.trade_log_entries.new
+    @log_entry.characters = [@character]
     authorize @log_entry
-    @magic_items    = @character.magic_items.where(trade_log_entry_id: nil)
-    @new_magic_item = @character.magic_items.build(trade_log_entry_id: 0)
   end
 
   def create
-    @new_magic_item = MagicItem.find_by_id(params[:trade_log_entry][:traded_magic_item]) || @character.magic_items.build(trade_log_entry_id: 0)
-    params[:trade_log_entry].delete(:traded_magic_item)
-
     @log_entry = @character.trade_log_entries.build(log_entries_params)
     @log_entry.traded_magic_item = @new_magic_item
+    @log_entry.characters = [@character]
 
     authorize @log_entry
-    @magic_items = @character.magic_items.where(trade_log_entry_id: [nil, @log_entry.id])
 
     if @log_entry.save
-      redirect_to user_character_path(current_user, @character, q: params[:q]), flash: { notice: 'Successfully created trade log entry' }
+      redirect_to user_character_path(current_user, @character, q: params[:q]),
+                  flash: { notice: 'Successfully created trade log entry' }
     else
-      flash.now[:error] = "Failed to create trade log entry: #{@log_entry.errors.full_messages.join(',')}"
+      flash.now[:error] = log_entry_error_message 'create'
       render :new, q: params[:q]
     end
   end
 
   def edit
     authorize @log_entry
-
-    @magic_items = @character.magic_items.where(trade_log_entry_id: [nil, @log_entry.id])
-    @new_magic_item = @log_entry.magic_items.last || @character.magic_items.build(trade_log_entry_id: 0)
   end
 
   def update
-    @new_magic_item = MagicItem.find_by_id(params[:trade_log_entry][:traded_magic_item]) || @character.magic_items.build(trade_log_entry_id: 0)
-    params[:trade_log_entry].delete(:traded_magic_item)
-
     authorize @log_entry
-    @magic_items = @character.magic_items.where(trade_log_entry_id: [nil, @log_entry.id])
-
     if @log_entry.update_attributes(log_entries_params)
-      redirect_to user_character_path(current_user, @character, q: params[:q]), flash: { notice: "Successfully updated trade log entry" }
+      redirect_to user_character_path(current_user, @character, q: params[:q]),
+                  flash: { notice: 'Successfully updated trade log entry' }
     else
-      flash.now[:error] = "Failed to update trade log entry: #{@log_entry.errors.full_messages.join(',')}"
+      flash.now[:error] = log_entry_error_message 'update'
       render :edit, q: params[:q]
     end
   end
@@ -68,14 +66,11 @@ class TradeLogEntriesController < AuthenticationController
     authorize @log_entry
     @log_entry.destroy
 
-    redirect_to user_character_path(current_user, @character, q: params[:q]), flash: { notice: "Successfully deleted trade log entry" }
+    redirect_to user_character_path(current_user, @character, q: params[:q]),
+                flash: { notice: 'Successfully deleted trade log entry' }
   end
 
   protected
-
-  def load_user
-    @user = User.find(params[:user_id])
-  end
 
   def load_character
     @character   = Character.find(params[:character_id])
@@ -85,7 +80,36 @@ class TradeLogEntriesController < AuthenticationController
     @log_entry   = LogEntry.find(params[:id])
   end
 
+  def load_current_magic_items
+    log_ids = [nil]
+    log_ids << @log_entry.id if @log_entry
+    @magic_items = @character.magic_items
+                             .where(trade_log_entry_id: log_ids)
+  end
+
+  def build_new_magic_item
+    magic_item = MagicItem.find_by(id: new_magic_item_params)
+    if magic_item.present?
+      @new_magic_item = magic_item
+    elsif @log_entry && @log_entry.magic_items.last
+      @new_magic_item = @log_entry.magic_items.last
+    else
+      @new_magic_item = MagicItem.new(trade_log_entry_id: 0)
+    end
+  end
+
+  def new_magic_item_params
+    params[:trade_log_entry][:traded_magic_item] if params[:trade_log_entry]
+  end
+
+  def clean_up_params
+    params[:trade_log_entry].delete(:traded_magic_item)
+  end
+
   def log_entries_params
-    params.require(:trade_log_entry).permit(:date_played, :downtime_gained, :traded_magic_item, :notes, magic_items_attributes: [:id, :name, :rarity, :notes, :_destroy])
+    params.require(:trade_log_entry)
+          .permit(:date_played, :downtime_gained, :traded_magic_item,
+                  :notes, magic_items_attributes: [:id, :name, :rarity,
+                                                   :notes, :_destroy])
   end
 end