+++ app/models/dm_log_entry.rb
@@ -1,5 +1,8 @@
+# :nodoc:
 class DmLogEntry < LogEntry
-  def is_dm_log_entry?
+  attr_accessor :character_id
+
+  def dm_log_entry?
     true
   end
 end