+++ app/models/log_assignment.rb
@@ -1,0 +1,6 @@
+class LogAssignment < ActiveRecord::Base
+  belongs_to :log_entry, required: true
+  belongs_to :character, required: true
+
+  has_one :user, through: :character
+end