+++ spec/features/dm_log_entries/delete_dm_log_entries_spec.rb
@@ -5,7 +5,8 @@ RSpec.feature 'DM Log Entries', type: :feature do
     @user = FactoryGirl.create(:user)
     login_as(@user, scope: :user)
     @character = FactoryGirl.create(:character, user: @user)
-    @dm_log_entry = FactoryGirl.create(:dm_log_entry, user: @user, character: @character)
+    @dm_log_entry = FactoryGirl.create(:dm_log_entry, user: @user)
+    @dm_log_entry.characters = [@character]
   end
 
   scenario 'Delete a DM Log Entry' do
@@ -17,7 +18,10 @@ RSpec.feature 'DM Log Entries', type: :feature do
 
     expect(DmLogEntry.count).to be(@dm_log_entry_count - 1)
 
-    expect(page).to_not have_text(@dm_log_entry.date_dmed.strftime('%Y-%m-%d %H:%M'))
+    visit user_dm_log_entries_path(@user)
+
+    expect(page).to_not have_text(@dm_log_entry.date_dmed
+                                               .strftime('%Y-%m-%d %H:%M'))
     expect(page).to_not have_text(@dm_log_entry.adventure_title)
     expect(page).to_not have_text(@dm_log_entry.session_num)
     expect(page).to_not have_text(@dm_log_entry.xp_gained)