+++ spec/features/dm_log_entries/show_dm_log_entries_spec.rb
@@ -4,8 +4,9 @@ RSpec.feature 'DM Log Entries', type: :feature do
   before(:each) do
     @user = FactoryGirl.create(:user)
     login_as(@user, scope: :user)
-    @character = FactoryGirl.create(:character, user: @user)
-    @dm_log_entry = FactoryGirl.create(:dm_log_entry, user: @user, character: @character)
+    @character    = FactoryGirl.create(:character, user: @user)
+    @dm_log_entry = FactoryGirl.create(:dm_log_entry, user: @user)
+    @dm_log_entry.characters = [@character]
   end
 
   scenario 'DM Log Entry index page should have information' do