+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 CHANGELOG
 ======
 
+v1.4.12
+---------------
+
+- Fix all characters selected when updating `Campaign log entry`.
+- Fix deleting campaign breaking access to `Campaign` page.
+- Fix sorting by characters in `DM Log Entries` index again.
+- Improve log entries code quality.
+
+v1.4.11
+---------------
+
+- Improve Campaigns#show and campaign log entry look.
+
+
+v1.4.10
+---------------
+
+- Change the way logs are assigned to allow for campaigns.
+- Add campaign log entries to `Campagins` feature.
+
 v1.4.9
 ---------------